debug: deep diagnostics for numtiles=0 + fix config defaults
The game now gets past Startup() but crashes after the window opens.
Root issues being investigated:

1. numtiles=0: loadpics reports no tiles loaded despite TILES000.ART
   being in the GRP. Added detailed logging to kopen4load, initgroupfile,
   and loadpics to trace exactly where the file lookup fails.

2. setgamemode(0,0,0): screen config not being read properly.
   - Fixed SCRIPT_GetNumber: long* vs int32* type mismatch (8 vs 4 bytes
     on 64-bit systems causes adjacent variable corruption)
   - Added screen defaults to CONFIG_SetDefaults() (ScreenMode=2,
     ScreenWidth=640, ScreenHeight=480)
   - Set FXDevice/MusicDevice=NumSoundCards in defaults (disables audio)

3. Extended startup logging past genspriteremaps/setbrightness to catch
   any post-setgamemode crash location.

4. initgroupfile now falls back to O_RDONLY if O_RDWR fails (Windows
   may have the GRP in a read-only location).

+++ compat/mact_stub.c
@@ -153,10 +153,10 @@ void SCRIPT_GetDoubleString(int handle, char *section, char *key, char *dest1, c
     }
 }
 
-int SCRIPT_GetNumber(int handle, char *section, char *key, long *dest) {
+int SCRIPT_GetNumber(int handle, char *section, char *key, int32_t *dest) {
     char buf[256];
     SCRIPT_GetString(handle, section, key, buf);
-    if (buf[0]) { *dest = atol(buf); return 1; }
+    if (buf[0]) { *dest = (int32_t)atol(buf); return 1; }
     *dest = 0;
     return 0;
 }
